+++ app/Enums/ECan.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Models\Enums;
+namespace App\Enums;
 
 enum ECan: string
 {
