+++ app/Traits/Models/HasPermissions.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace App\Models\Traits;
+namespace App\Traits\Models;
 
-use App\Models\Enums\ECan;
+use App\Enums\ECan;
 use App\Models\Permission;
+use Illuminate\Support\Facades\Cache;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
-use Illuminate\Support\Facades\Cache;
 
 trait HasPermissions
 {
@@ -27,7 +27,7 @@ trait HasPermissions
 
         $cacheKey = $this->getPermissionsCacheKey();
         Cache::forget($cacheKey);
-        Cache::rememberForever($cacheKey, fn () => $this->permissions);
+        Cache::rememberForever($cacheKey, fn() => $this->permissions);
     }
 
     public function hasPermissionTo(ECan|string $key): bool
